refactor: dp 사용 및 가독성 개선
