--- src/main/java/com/daemawiki/archive/daemawiki/domain/mail/repository/AuthUserRepository.java
@@ -1,12 +1,0 @@
-package com.daemawiki.archive.daemawiki.domain.mail.repository;
-
-import reactor.core.publisher.Mono;
-
-public interface AuthUserRepository {
-    Mono<Boolean> save(String mail);
-
-    Mono<Boolean> existsByEmail(String mail);
-
-    Mono<Void> delete(String mail);
-
-}
