+++ src/main/java/com/daemawiki/internal/core/usecase/mail/MailVerifyUseCase.java
@@ -1,5 +1,6 @@
 package com.daemawiki.internal.core.usecase.mail;
 
+import com.daemawiki.internal.core.domain.model.primitive.mail.AuthCode;
 import com.daemawiki.internal.core.domain.model.primitive.user.personal.Email;
 import reactor.core.publisher.Mono;
 
@@ -7,7 +8,7 @@ public interface MailVerifyUseCase {
 
     Mono<Void> verify(
             Email target,
-            String code
+            AuthCode authCode
     );
 
 }
