+++ src/main/java/com/daemawiki/internal/data/repository/mail/AuthCodeRepository.java
@@ -1,4 +1,4 @@
-package com.daemawiki.archive.daemawiki.domain.mail.repository;
+package com.daemawiki.internal.data.repository.mail;
 
 import com.daemawiki.internal.core.domain.model.dto.mail.AuthCodeDTO;
 import com.daemawiki.internal.core.domain.model.primitive.user.personal.Email;
@@ -9,8 +9,6 @@ public interface AuthCodeRepository {
     Mono<Boolean> save(AuthCodeDTO model);
 
     Mono<AuthCodeDTO> findByMail(Email email);
-
-    Mono<Void> delete(AuthCodeDTO model);
 
     Mono<Long> deleteByEmail(Email email);
