+++ src/main/java/com/daemawiki/internal/data/repository/mail/impl/AuthMailRepositoryImpl.java
@@ -1,7 +1,8 @@
-package com.daemawiki.archive.daemawiki.domain.mail.repository.impl;
+package com.daemawiki.internal.data.repository.mail.impl;
 
 import com.daemawiki.external.exception.custom.CustomExceptionFactory;
-import com.daemawiki.archive.daemawiki.domain.mail.repository.AuthUserRepository;
+import com.daemawiki.internal.core.domain.model.primitive.user.personal.Email;
+import com.daemawiki.internal.data.repository.mail.AuthMailRepository;
 import com.daemawiki.external.database.redis.RedisKey;
 import com.daemawiki.external.database.redis.storage.RedisOperation;
 import lombok.RequiredArgsConstructor;
@@ -14,14 +15,14 @@ import java.time.Duration;
 
 @Repository
 @RequiredArgsConstructor
-@Slf4j(topic = "인증된 유저 레디스 레포지토리")
-class AuthUserRepositoryImpl implements AuthUserRepository {
+@Slf4j(topic = "AuthMailRepositoryImpl")
+class AuthMailRepositoryImpl implements AuthMailRepository {
     private static final String AUTH_MAIL = RedisKey.AUTH_USER.getKey();
 
-    private final RedisOperation redisOperation;
+    private final RedisOperation<Email> redisOperation;
 
     @Override
-    public Mono<Boolean> save(String mail) {
+    public Mono<Boolean> save(final Email mail) {
         return handleError(redisOperation.save(
                 AUTH_MAIL + mail,
                 mail,
@@ -30,16 +31,16 @@ class AuthUserRepositoryImpl implements AuthUserRepository {
     }
 
     @Override
-    public Mono<Boolean> existsByEmail(String mail) {
+    public Mono<Boolean> existsByEmail(final Email mail) {
         return handleError(redisOperation.hasKey(AUTH_MAIL + mail));
     }
 
     @Override
-    public Mono<Void> delete(String mail) {
+    public Mono<Void> delete(final Email mail) {
         return handleError(redisOperation.delete(AUTH_MAIL + mail).then());
     }
 
-    private static <T> Mono<T> handleError(Mono<T> mono) {
+    private static <T> Mono<T> handleError(final Mono<T> mono) {
         return mono.onErrorMap(e -> {
             log.error("#- Error: " + e);
             return e instanceof RedisConnectionFailureException ?
